third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/registries/element-mutation-null-registry-removal-expected.txt

This is a testharness.js-based test.
[FAIL] An element's null customElementRegistry should not mutate after removal from a declarative shadow root with shadowrootcustomelementregistry.
  assert_equals: customElementRegistry should be null while in null-registry shadow root expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
